Fix the async problem
Requests have to be async, so when we loop through the nodes and call a request for each node, they all return results at different times... aka unordered... and we need our results to be ordered. So I'm using a global variable results, it gets reset every runScrapey call, and the first element(results[0]) gets set to the amount of requests we are going to send, that way we can check to see if we are the last request and to return the results if so.